feat: finish decomposition of useForm hook
